capture/run cycle is working! fixed a memory problem where limb was access outside image

+++ sonic-gesture/src/matcher.h
@@ -13,7 +13,7 @@ class Matcher {
     Mat handimg;
     Mat train;
     Mat labels_mat;
-    vector<int> labels;
+    vector<float> labels;
 
 public:
     Matcher();
